chore: changed file names add split up some components
