+++ components/auth/Profile.tsx
@@ -2,13 +2,13 @@
 
 import React from "react";
 import { useAppAuthStore } from "../../store/auth/app-auth-store";
-import Link from "next/link";
 import { Button } from "../ui/button";
 
 import { motion } from "framer-motion";
 import { UserDropdown } from "../navbar/user-dropdown/UserDropdown";
 import { useClerk } from "@clerk/nextjs";
 import { handleSignOutWithToast } from "@/store/auth/AuthStoreProvider";
+import { AuthLinks } from "./AuthLinks";
 
 type Props = object & {};
 
@@ -35,24 +35,7 @@ export const Profile: React.FC<Props> = () => {
     }
 
     if (authState === "guest") {
-        return (
-            <>
-                <Link
-                    href="/sign-up"
-                    className="link__with__button ring-slate-600"
-                >
-                    <Button role="link" tabIndex={-1} variant={"outline"}>
-                        Sign up
-                    </Button>
-                </Link>
-
-                <Link href="/sign-in" className="link__with__button ring-white">
-                    <Button role="link" tabIndex={-1}>
-                        Sign in
-                    </Button>
-                </Link>
-            </>
-        );
+        return <AuthLinks />;
     }
 
     if (!user?.username) {
